requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
pycparser==2.21
PyNaCl==1.5.0
pyTelegramBotAPI==4.14.1
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
yarl==1.9.4
